MB-10533 Move request-related class + funcs to utils/request.py
They don't really belong in hosts.py. Not sure if this is the best named file...
